minor changes from int to int16_t, and revision of makefile to use old C libraries and include files to reduce size of file

+++ zas/kwd.h
@@ -30,8 +30,8 @@
  */
 typedef struct _kwd {
     char *name;
-    int id;
-    int flags;
+    int16_t id;
+    int16_t flags;
 } kwd_t;
 
 #define G_SIMPLE   1
